Prototype-1/requirements.txt
streamlit>=1.28.0
openai-whisper>=20230918
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
google-generativeai>=0.3.0